+++ services/api-server/src/simcore_service_api_server/services/_service_exception_handling.py
@@ -1,5 +1,6 @@
 import logging
 from contextlib import contextmanager
+from functools import wraps
 from typing import Mapping
 
 import httpx
@@ -8,6 +9,20 @@ from pydantic import ValidationError
 from servicelib.error_codes import create_error_code
 
 _logger = logging.getLogger(__name__)
+
+
+def service_status_mapper(
+    service_name: str, http_status_map: Mapping[int, tuple[int, str | None]]
+):
+    def decorator(func):
+        @wraps(func)
+        async def wrapper(*args, **kwargs):
+            with backend_service_exception_handler(service_name, http_status_map):
+                return await func(*args, **kwargs)
+
+        return wrapper
+
+    return decorator
 
 
 @contextmanager
@@ -19,7 +34,8 @@ def backend_service_exception_handler(
     except ValidationError as exc:
         error_code = create_error_code(exc)
         _logger.exception(
-            "Invalid data exchanged with webserver service [%s]",
+            "Invalid data exchanged with %s service [%s] ",
+            service_name,
             error_code,
             extra={"error_code": error_code},
         )
